refactor(EE10.03): extract runtime-config.ts from orchestrator.ts
Move _config singleton and all config init/access/resolution functions
into tools/delivery/runtime-config.ts. Migrate config tests verbatim
into test/runtime-config.test.ts. orchestrator.ts re-exports all config
surface unchanged.
