+++ tools/delivery/test/orchestrator.test.ts
@@ -30,7 +30,6 @@ import {
   formatNotificationMessage,
   formatReviewWindowMessage,
   generateRunDeliverInvocation,
-  inferPackageManager,
   initOrchestratorConfig,
   VALID_REVIEW_POLICY_STAGE_VALUES,
   recordCodexPreflight,
@@ -56,7 +55,6 @@ import {
   shouldAutoRecordReviewSkippedForPollReview,
   summarizeStateDifferences,
   syncStateFromExisting,
-  syncStateFromScratch,
   runProcessResult,
   formatCurrentTicketStatus,
   formatAdvanceBoundaryGuidance,
@@ -3860,287 +3858,6 @@ describe('delivery orchestrator', () => {
         process.env.AI_CODE_REVIEW_TRIAGER = original;
       }
     }
-  });
-
-  describe('orchestrator config', () => {
-    it('returns defaults when config file is absent', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        const config = await loadOrchestratorConfig(tempDir);
-        expect(config).toEqual({});
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('loads a partial config and preserves specified fields', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ defaultBranch: 'develop', runtime: 'node' }),
-        );
-
-        const config = await loadOrchestratorConfig(tempDir);
-        expect(config).toEqual({
-          defaultBranch: 'develop',
-          planRoot: undefined,
-          runtime: 'node',
-          packageManager: undefined,
-          ticketBoundaryMode: undefined,
-        });
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws on invalid runtime value', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ runtime: 'deno' }),
-        );
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /Invalid runtime "deno"/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws on invalid packageManager value', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ packageManager: 'cargo' }),
-        );
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /Invalid packageManager "cargo"/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws on invalid ticketBoundaryMode value', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ ticketBoundaryMode: 'sprint' }),
-        );
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /Invalid ticketBoundaryMode "sprint"/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws on non-string defaultBranch', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ defaultBranch: 42 }),
-        );
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /Invalid defaultBranch/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws when config json is not an object', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(join(tempDir, 'orchestrator.config.json'), '[]');
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /must contain a JSON object/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws on blank defaultBranch', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ defaultBranch: '   ' }),
-        );
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /Expected a non-blank string/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('throws on blank planRoot', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-'));
-      try {
-        await writeFile(
-          join(tempDir, 'orchestrator.config.json'),
-          JSON.stringify({ planRoot: '   ' }),
-        );
-
-        await expect(loadOrchestratorConfig(tempDir)).rejects.toThrow(
-          /Expected a non-blank string/,
-        );
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('resolves empty config to defaults', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-resolve-'));
-      try {
-        const resolved = resolveOrchestratorConfig({}, tempDir);
-        expect(resolved).toEqual({
-          defaultBranch: 'main',
-          planRoot: 'docs',
-          runtime: 'bun',
-          packageManager: 'npm',
-          ticketBoundaryMode: 'cook',
-          reviewPolicy: {
-            selfAudit: 'skip_doc_only',
-            codexPreflight: 'skip_doc_only',
-            externalReview: 'skip_doc_only',
-          },
-        });
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('merges partial config with defaults', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-resolve-'));
-      try {
-        const resolved = resolveOrchestratorConfig(
-          { defaultBranch: 'develop', planRoot: 'specifications' },
-          tempDir,
-        );
-        expect(resolved.defaultBranch).toBe('develop');
-        expect(resolved.planRoot).toBe('specifications');
-        expect(resolved.runtime).toBe('bun');
-        expect(resolved.packageManager).toBe('npm');
-        expect(resolved.ticketBoundaryMode).toBe('cook');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('preserves configured ticketBoundaryMode when resolving config', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-cfg-resolve-'));
-      try {
-        const resolved = resolveOrchestratorConfig(
-          { ticketBoundaryMode: 'gated' },
-          tempDir,
-        );
-        expect(resolved.ticketBoundaryMode).toBe('gated');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('infers bun from bun.lock', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-pm-'));
-      try {
-        await writeFile(join(tempDir, 'bun.lock'), '');
-        expect(inferPackageManager(tempDir)).toBe('bun');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('infers pnpm from pnpm-lock.yaml', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-pm-'));
-      try {
-        await writeFile(join(tempDir, 'pnpm-lock.yaml'), '');
-        expect(inferPackageManager(tempDir)).toBe('pnpm');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('infers yarn from yarn.lock', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-pm-'));
-      try {
-        await writeFile(join(tempDir, 'yarn.lock'), '');
-        expect(inferPackageManager(tempDir)).toBe('yarn');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('infers npm from package-lock.json', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-pm-'));
-      try {
-        await writeFile(join(tempDir, 'package-lock.json'), '{}');
-        expect(inferPackageManager(tempDir)).toBe('npm');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('falls back to npm when no lockfile is present', async () => {
-      const tempDir = await mkdtemp(join(tmpdir(), 'orch-pm-'));
-      try {
-        expect(inferPackageManager(tempDir)).toBe('npm');
-      } finally {
-        await rm(tempDir, { recursive: true });
-      }
-    });
-
-    it('syncStateFromScratch uses configured defaultBranch for first ticket baseBranch', () => {
-      initOrchestratorConfig({
-        defaultBranch: 'develop',
-        planRoot: 'docs',
-        runtime: 'bun',
-        packageManager: 'bun',
-        ticketBoundaryMode: 'cook',
-      });
-
-      try {
-        const options = createOptions({
-          planPath: 'docs/02-delivery/phase-03/implementation-plan.md',
-        });
-
-        const synced = syncStateFromScratch(
-          [
-            {
-              id: 'P3.01',
-              title: 'First Ticket',
-              slug: 'first-ticket',
-              ticketFile: 'docs/02-delivery/phase-03/ticket-01-first-ticket.md',
-            },
-          ],
-          '/workspace/test',
-          options,
-        );
-
-        expect(synced.tickets[0]?.baseBranch).toBe('develop');
-      } finally {
-        initOrchestratorConfig({
-          defaultBranch: 'main',
-          planRoot: 'docs',
-          runtime: 'bun',
-          packageManager: 'bun',
-          ticketBoundaryMode: 'cook',
-        });
-      }
-    });
   });
 
   it('renders npm deliver invocations with a separator', () => {
